# docs/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Hamlaneh API
  description: >
    Self-hosted team communication platform API. Phase 1.1 surface: health,
    authentication, current user, and the admin user-management endpoints that
    back the dashboard (public registration is off by default; users are
    created from the admin dashboard or via invite links). Phase 1.2 adds the
    messaging surface the chat shell needs: channels and 1:1 direct messages
    with membership, cursor-paged message history with idempotent send, edit
    and soft delete, per-user read positions, message search, and the
    WebSocket upgrade that carries realtime delivery. Phase 1.1b completes the
    identity surface: email password reset, TOTP two-step verification with
    recovery codes, and self-service session management.
  version: 0.6.0
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: /
    description: Same-origin deployment behind Caddy.
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [auth]
      operationId: login
      summary: Authenticate with username/email and password.
      description: >
        Rate-limited. On success sets the session and refresh cookies and
        returns the authenticated user. Responses for unknown user and wrong
        password are identical (no account enumeration).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authenticated; session cookies set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "202":
          description: >
            Password verified, but two-step verification is on: no session
            exists yet. The challenge cookie is set; complete sign-in at
            /api/v1/auth/login/totp. A distinct status so no client can
            mistake this for a signed-in state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TwoFactorChallenge"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/auth/login/totp:
    post:
      tags: [auth]
      operationId: completeTotpLogin
      summary: Complete a two-step sign-in with an authenticator or recovery code.
      description: >
        The second half of login. Authenticated by the hamlaneh_2fa challenge
        cookie the 202 login set — deliberately no security scheme here,
        mirroring /auth/refresh; the authz matrix classifies it as
        challenge-cookie-gated. `code` is the six-digit authenticator code or
        one unused recovery code (case-insensitive, hyphens optional); a
        recovery code is consumed by the attempt that uses it. An accepted
        authenticator code is never accepted twice, even inside its
        thirty-second window. Five wrong codes revoke the challenge and the
        caller starts again at the password step. On success the challenge is
        consumed and session cookies are set exactly as a plain login sets them.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TotpLoginRequest"
      responses:
        "200":
          description: Authenticated; session cookies set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: >
            No live challenge — missing, expired, consumed or revoked (code
            not_authenticated; return to the password step) — or a live
            challenge with a wrong code (code invalid_totp_code; the cells
            clear and the caller retries).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/auth/logout:
    post:
      tags: [auth]
      operationId: logout
      summary: End the current session and clear cookies.
      security:
        - sessionCookie: []
      responses:
        "204":
          description: Session revoked; cookies cleared.
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/refresh:
    post:
      tags: [auth]
      operationId: refreshSession
      summary: Rotate the refresh token and mint a new session.
      description: >
        Presenting an already-rotated refresh token revokes the whole token
        family (reuse detection) and returns 401.
      responses:
        "204":
          description: New session cookies set.
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/change-password:
    post:
      tags: [auth]
      operationId: changePassword
      summary: Change the authenticated user's password.
      description: >
        Requires the current password. Allowed (and required) while
        must_change_password is set. On success all OTHER sessions of the user
        are revoked; the current session stays valid and the flag clears.
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "204":
          description: Password changed; other sessions revoked.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Current password incorrect (code invalid_current_password).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/auth/reset-request:
    post:
      tags: [auth]
      operationId: requestPasswordReset
      summary: Request a password-reset email.
      description: >
        Enumeration-safe: 202 with an empty body whether or not the address
        matches an account, the token mint and hash run either way so timing
        does not distinguish them, and the mail is dispatched asynchronously
        so SMTP latency never sits on the response. Tokens are 256-bit,
        single-use, thirty-minute, stored as SHA-256; issuing a new one
        invalidates the account's previous unused tokens, so at most one link
        is ever live. Rate-limited per IP and per PRESENTED address — keyed to
        the string as typed, existing or not, because a limiter that only
        counted real accounts would itself leak which addresses exist.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetRequest"
      responses:
        "202":
          description: >
            Accepted. If that address belongs to an account a reset link is on
            its way — the response is identical either way.
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/auth/reset-complete:
    post:
      tags: [auth]
      operationId: completePasswordReset
      summary: Set a new password with an emailed reset token.
      description: >
        Consumes the token and stores the new password in one transaction.
        Unknown, expired and already-used tokens answer identically so a
        replayed link learns nothing. Completing a reset revokes EVERY session
        family — unlike change-password, which keeps the current one, because
        a reset means the old password may be in someone else's hands — clears
        must_change_password, and consumes the account's other outstanding
        reset tokens and pending two-step challenges. It does NOT disable
        two-step verification: the token proves control of the mailbox, not of
        the authenticator. No cookies are set; the user signs in fresh.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetCompleteRequest"
      responses:
        "204":
          description: Password replaced; every session family revoked.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: >
            Token unknown, expired or already used (code invalid_reset_token)
            — one answer for all three.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/users/me:
    get:
      tags: [users]
      operationId: getCurrentUser
      summary: The authenticated user.
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Current user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"

    patch:
      tags: [users]
      operationId: updateCurrentUser
      summary: Change your own account settings.
      description: >
        Locale belongs on the account rather than only in the browser: a
        person who reads Persian reads it on their phone too, and a
        preference that lives in localStorage silently reverts on every new
        device and every cleared browser. The client still applies its local
        choice immediately — the round trip must not make the interface
        wait — and this is what makes the choice follow the person.

        Unlike its siblings this stays open while must_change_password is
        set. The forced-change screen renders the language switcher, so
        refusing the save there would leave a control that appears to work
        and silently does not; and choosing which language to read the
        password rules in is not something the gate exists to prevent.
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCurrentUserRequest"
      responses:
        "200":
          description: The updated account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/users/me/totp:
    get:
      tags: [users]
      operationId: getTotpStatus
      summary: The caller's two-step verification state (the Security card).
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Two-step state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TotpStatus"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/users/me/totp/setup:
    post:
      tags: [users]
      operationId: startTotpSetup
      summary: Start (or restart) two-step verification setup.
      description: >
        Generates a fresh 160-bit secret and returns the manual key, the
        otpauth URI and a server-rendered QR (the settings handoff: rendered
        server-side from the secret, never a stored image). The pending setup
        lives one hour; calling again before activation replaces it, which is
        also why Cancel needs no endpoint. The account stays password-only
        until activate. Answers carry Cache-Control: no-store.
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Pending setup created or replaced.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TotpSetup"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Two-step verification is already on (code totp_already_enabled).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/users/me/totp/verify:
    post:
      tags: [users]
      operationId: verifyTotpSetup
      summary: Verify the first authenticator code (setup step 2).
      description: >
        Checks a six-digit code against the pending secret. A wrong code does
        not restart setup — the secret stays valid and the client clears the
        cells — but five wrong codes revoke the pending setup, because an
        uncapped verifier is a brute-force oracle. On success the ten recovery
        codes are generated and returned, shown exactly once (only argon2id
        hashes are kept). Two-step verification is still OFF until activate.
        Answers carry Cache-Control: no-store.
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyTotpSetupRequest"
      responses:
        "200":
          description: Code accepted; the recovery codes for step 3.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecoveryCodes"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Wrong code (code invalid_totp_code); the setup survives.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: >
            No pending setup, it expired, or the attempt cap revoked it (code
            totp_setup_expired) — start again at step 1.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/users/me/totp/activate:
    post:
      tags: [users]
      operationId: activateTotp
      summary: Turn two-step verification on (setup step 3).
      description: >
        The action behind the saved-the-codes acknowledgement. Only a VERIFIED
        pending setup activates: the server never turns two-step verification
        on for an account that has not proven its authenticator and been shown
        its recovery codes, so closing the panel at step 3 can never strand an
        account whose only fallback was never displayed.
      security:
        - sessionCookie: []
      responses:
        "204":
          description: Two-step verification is on for the next sign-in.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: >
            Nothing to activate — no pending setup, not yet verified, or
            expired (code totp_setup_not_verified).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v1/users/me/sessions/revoke-others:
    post:
      tags: [users]
      operationId: revokeOtherSessions
      summary: Sign out everywhere else.
      description: >
        Revokes every live family except the caller's own. 204 even when
        nothing else was signed in; the client already knows the count.
      security:
        - sessionCookie: []
      responses:
        "204":
          description: All other families revoked.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: hamlaneh_session
      description: >
        HttpOnly+Secure+SameSite=Strict session cookie. Mutating requests must
        also send the X-Hamlaneh-CSRF header.
...
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    LoginRequest:
      type: object
      required: [identifier, password]
      properties:
        identifier:
          type: string
          description: Username or email.
          minLength: 1
          maxLength: 320
        password:
          type: string
          format: password
          minLength: 1
          maxLength: 1024
...
    PasswordResetRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          maxLength: 320
...